pyproject.toml
[tool.poetry]
name = "crypto-exchange-dashboard"
version = "0.1.0"
description = "Crtptho exchages dashboard"
authors = ["tiziaco <tiziano.iaco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.1"
ccxt = "^4.1.40"
pandas = "^2.1.2"
numpy = "^1.26.1"
flask-sqlalchemy = "^3.1.1"
python-dotenv = "^1.0.0"
flask-migrate = "^4.0.5"
dash = "^2.14.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
